complete move to subroutines system
