+++ src/entry.ts
@@ -2,6 +2,11 @@ import 'dotenv/config';
 import { app, prisma, minuteInterval, hourInterval } from './app.js';
 import "./subroutines/onboarding.js";
 import "./subroutines/hackhour.js";
+import "./subroutines/goals.js";
+import "./subroutines/picnics.js";
+import "./subroutines/misc.js";
+
+import { Constants } from './constants.js';
 
 const mainLoop = async () => {
     await prisma.$connect();
@@ -10,12 +15,6 @@ const mainLoop = async () => {
     hourInterval.start();    
 
     console.log(`⏳ Let the Hack Houring Begin! Running on port ${process.env.PORT || 3000}...`);
-
-    await app.client.chat.postMessage({
-        token: process.env.SLACK_BOT_TOKEN,        
-        channel: process.env.LOG_CHANNEL || 'C0P5NE354' ,
-        text: `⏳ Let the Hack Houring Begin!`
-    });    
 }
 
 mainLoop().catch(async (error) => {
