+++ src/subroutines/picnics.ts
@@ -1,0 +1,7 @@
+import { app, prisma } from '../app.js';
+
+import { Callbacks } from '../views/picnics.js';
+
+app.view(Callbacks.PICNIC, async ({ ack, body, client }) => {
+    await ack();
+});
