+++ src/utils/intervalManager.ts
@@ -4,6 +4,7 @@
 export class IntervalManager {
     private interval: number;
     private callbacks: (() => void)[] = [];    
+    private delay: number = 0;
     
     public constructor(interval: number) {
         this.interval = interval;
@@ -14,8 +15,14 @@ export class IntervalManager {
     }
 
     public start(): void {
-        setInterval(() => {
-            this.callbacks.forEach(callback => callback());
-        }, this.interval);
+        setTimeout(() => {
+            setInterval(() => {
+                this.callbacks.forEach(callback => callback());
+            }, this.interval);
+        }, this.delay);
+    }
+
+    public setDelay(delay: number): void {
+        this.delay = delay;
     }
 }
