+++ src/views/goals.ts
@@ -1,20 +1,46 @@
 import { prisma } from "../app.js";
 import { View } from "@slack/bolt";
+import { formatHour } from "../utils/string.js";
 
 export const Callbacks = {
     GOALS: 'goals-callback',
+    CREATE: 'createGoal-callback',
+    DELETE: 'deleteGoal-callback',
+    ERROR: 'goals-error'
+}
+
+export const Actions = {
+    SELECT: 'selectGoal',
+    CREATE: 'createGoal',
+    DELETE: 'deleteGoal'
 }
 
 export class Views {
     public static async goals(userId: string): Promise<View> {
-        const goals = await prisma.goal.findMany({
+        const userData = await prisma.user.findUnique({
             where: {
-                userId
+                slackId: userId
+            }
+        });
+
+        if (!userData) {
+            throw new Error(`User ${userId} not found.`);
+        }
+
+        const goals = await prisma.goals.findMany({
+            where: {
+                slackId: userId
+            }
+        });
+
+        const selectedGoal = await prisma.goals.findFirst({
+            where: {
+                goalId: userData.selectedGoal
             }
         });
 
         if (goals.length === 0) {
-            throw new Error("No goals found.");
+            throw new Error(`Goals of ${userId} not found.`);
         }
 
         return {
@@ -22,7 +48,7 @@ export class Views {
             "type": "modal",
             "submit": {
                 "type": "plain_text",
-                "text": "I'm done",
+                "text": "Select",
                 "emoji": true
             },
             "close": {
@@ -45,21 +71,29 @@ export class Views {
                     }
                 },
                 {
-                    "type": "actions",
+                    "type": "actions",                    
                     "elements": [
                         {
                             "type": "radio_buttons",
+                            "initial_option": {
+                                "text": {
+                                    "type": "plain_text",
+                                    "text": `${selectedGoal?.goalName}`,
+                                    "emoji": true
+                                },
+                                "value": `${selectedGoal?.goalId}`
+                            },
                             "options": goals.map(goal => {
                                 return {
                                     "text": {
                                         "type": "plain_text",
-                                        "text": "*plain_text option 0*",
-                                        "emoji": true
+                                        "text": `${goal.goalName}`,
+                                        "emoji": true,
                                     },
                                     "value": goal.goalId
                                 }
                             }),
-                            "action_id": "selectGoal"
+                            "action_id": Actions.SELECT
                         }
                     ]
                 },
@@ -73,7 +107,8 @@ export class Views {
                                 "text": "Create Goal",
                                 "emoji": true
                             },
-                            "value": "create"
+                            "value": Actions.CREATE,
+                            "action_id": Actions.CREATE
                         },
                         {
                             "type": "button",
@@ -82,7 +117,8 @@ export class Views {
                                 "text": "Delete Goal",
                                 "emoji": true
                             },
-                            "value": "delete"
+                            "value": Actions.DELETE,
+                            "action_id": Actions.DELETE
                         }
                     ]
                 },
@@ -93,7 +129,113 @@ export class Views {
                     "type": "section",
                     "text": {
                         "type": "mrkdwn",
-                        "text": "You've spent *{hours}* hours working on _{goal}_."
+                        "text": `You've spent *${formatHour(selectedGoal?.minutes)}* hours working on _${selectedGoal?.goalName}_.`
+                    }
+                }
+            ]
+        }
+    }
+
+    public static createGoal(): View {
+        return {
+            "type": "modal",
+            "callback_id": Callbacks.CREATE,
+            "submit": {
+                "type": "plain_text",
+                "text": "Submit",
+                "emoji": true
+            },
+            "close": {
+                "type": "plain_text",
+                "text": "Cancel",
+                "emoji": true
+            },
+            "title": {
+                "type": "plain_text",
+                "text": "Goals",
+                "emoji": true
+            },
+            "blocks": [
+                {
+                    "type": "input",
+                    "element": {
+                        "type": "plain_text_input",
+                        "action_id": "goalName"
+                    },
+                    "label": {
+                        "type": "plain_text",
+                        "text": "Enter Goal Name:",
+                        "emoji": true				
+                    },
+                    "block_id": "goal"
+                }
+            ]
+        }
+    }
+
+    public static deleteGoal(): View {
+        return {
+            "type": "modal",
+            "callback_id": Callbacks.DELETE,
+            "submit": {
+                "type": "plain_text",
+                "text": "Yes",
+                "emoji": true
+            },
+            "close": {
+                "type": "plain_text",
+                "text": "No",
+                "emoji": true
+            },
+            "title": {
+                "type": "plain_text",
+                "text": "Goals",
+                "emoji": true
+            },
+            "blocks": [
+                {
+                    "type": "rich_text",
+                    "elements": [
+                        {
+                            "type": "rich_text_section",
+                            "elements": [
+                                {
+                                    "type": "text",
+                                    "text": "Are you sure you want delete this goal?"
+                                }
+                            ]
+                        }
+                    ]
+                }
+            ]
+        };
+    }
+
+    public static error(error: string): View {
+        return {
+            "type": "modal",
+            "callback_id": Callbacks.ERROR,
+            "submit": {
+                "type": "plain_text",
+                "text": "Okay",
+                "emoji": true
+            },
+            "close": {
+                "type": "plain_text",
+                "text": "Close",
+                "emoji": true
+            },
+            "title": {
+                "type": "plain_text",
+                "text": "Goals",
+                "emoji": true
+            },
+            "blocks": [
+                {
+                    "type": "section",
+                    "text": {
+                        "type": "mrkdwn",
+                        "text": error
                     }
                 }
             ]
